+++ Models/Producto.cs
@@ -2,10 +2,10 @@ public class Producto
 {
     private int idProducto;
     private string descripcion;
-    private int precio;
+    private decimal precio;
 
     // PROPERTIES
     public int IdProducto { get => idProducto; set => idProducto = value; }
     public string Descripcion { get => descripcion; set => descripcion = value; }
-    public int Precio { get => precio; set => precio = value; }
+    public decimal Precio { get => precio; set => precio = value; }
 }
